# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ArtworkStatus = (typeof ArtworkStatus)[keyof typeof ArtworkStatus];

export const ArtworkStatus = {
  available: "available",
  sold: "sold",
  unavailable: "unavailable",
} as const;

export interface Artwork {
  id: number;
  slug: string;
  title: string;
  /** @nullable */
  medium?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  price?: number | null;
  status: ArtworkStatus;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  isFeatured: boolean;
  /** @nullable */
  year?: number | null;
  createdAt: string;
}

export type CreateCheckoutSessionBodyPurchaseType =
  (typeof CreateCheckoutSessionBodyPurchaseType)[keyof typeof CreateCheckoutSessionBodyPurchaseType];

export const CreateCheckoutSessionBodyPurchaseType = {
  original: "original",
  print: "print",
} as const;

export interface CreateCheckoutSessionBody {
  artworkSlug: string;
  purchaseType: CreateCheckoutSessionBodyPurchaseType;
  /** @nullable */
  customerEmail?: string | null;
  successUrl: string;
  cancelUrl: string;
}

export interface CheckoutSessionResponse {
  url: string;
  sessionId: string;
}

export interface VerifyCheckoutBody {
  sessionId: string;
}

export type VerifyCheckoutResponsePurchaseType =
  (typeof VerifyCheckoutResponsePurchaseType)[keyof typeof VerifyCheckoutResponsePurchaseType];

export const VerifyCheckoutResponsePurchaseType = {
  original: "original",
  print: "print",
} as const;

export interface VerifyCheckoutResponse {
  success: boolean;
  purchaseType: VerifyCheckoutResponsePurchaseType;
  artworkTitle: string;
  message: string;
}

export interface ErrorResponse {
  error: string;
}