# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: artworks
    description: Artwork operations
  - name: checkout
    description: Checkout and payment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /artworks:
    get:
      operationId: listArtworks
      tags: [artworks]
      summary: List all artworks
      description: Returns all artworks ordered with featured pieces first
      responses:
        "200":
          description: List of artworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artwork"

  /artworks/{slug}:
    get:
      operationId: getArtwork
      tags: [artworks]
      summary: Get artwork by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artwork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Artwork not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout/verify:
    post:
      operationId: verifyCheckout
      tags: [checkout]
      summary: Verify a completed Stripe checkout session and fulfill the order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCheckoutBody"
      responses:
        "200":
          description: Order verified and fulfilled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyCheckoutResponse"
        "400":
          description: Payment not completed or invalid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Artwork:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        medium:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        price:
          type: ["integer", "null"]
        status:
          type: string
          enum: [available, sold, unavailable]
        description:
          type: ["string", "null"]
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        year:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - status
        - imageUrl
        - isFeatured
        - createdAt

    CreateCheckoutSessionBody:
      type: object
      properties:
        artworkSlug:
          type: string
        purchaseType:
          type: string
          enum: [original, print]
        customerEmail:
          type: ["string", "null"]
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - artworkSlug
        - purchaseType
        - successUrl
        - cancelUrl

    CheckoutSessionResponse:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required:
        - url
        - sessionId

    VerifyCheckoutBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId

    VerifyCheckoutResponse:
      type: object
      properties:
        success:
          type: boolean
        purchaseType:
          type: string
          enum: [original, print]
        artworkTitle:
          type: string
        message:
          type: string
      required:
        - success
        - purchaseType
        - artworkTitle
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error