# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all artworks ordered with featured pieces first
 * @summary List all artworks
 */
export const ListArtworksResponseItem = zod.object({
  id: zod.number(),
  slug: zod.string(),
  title: zod.string(),
  medium: zod.string().nullish(),
  dimensions: zod.string().nullish(),
  price: zod.number().nullish(),
  status: zod.enum(["available", "sold", "unavailable"]),
  description: zod.string().nullish(),
  imageUrl: zod.string(),
  isFeatured: zod.boolean(),
  year: zod.number().nullish(),
  createdAt: zod.date(),
});
export const ListArtworksResponse = zod.array(ListArtworksResponseItem);

/**
 * @summary Get artwork by slug
 */
export const GetArtworkParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetArtworkResponse = zod.object({
  id: zod.number(),
  slug: zod.string(),
  title: zod.string(),
  medium: zod.string().nullish(),
  dimensions: zod.string().nullish(),
  price: zod.number().nullish(),
  status: zod.enum(["available", "sold", "unavailable"]),
  description: zod.string().nullish(),
  imageUrl: zod.string(),
  isFeatured: zod.boolean(),
  year: zod.number().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary Create a Stripe checkout session
 */
export const CreateCheckoutSessionBody = zod.object({
  artworkSlug: zod.string(),
  purchaseType: zod.enum(["original", "print"]),
  customerEmail: zod.string().nullish(),
  successUrl: zod.string(),
  cancelUrl: zod.string(),
});

export const CreateCheckoutSessionResponse = zod.object({
  url: zod.string(),
  sessionId: zod.string(),
});

/**
 * @summary Verify a completed Stripe checkout session and fulfill the order
 */
export const VerifyCheckoutBody = zod.object({
  sessionId: zod.string(),
});

export const VerifyCheckoutResponse = zod.object({
  success: zod.boolean(),
  purchaseType: zod.enum(["original", "print"]),
  artworkTitle: zod.string(),
  message: zod.string(),
});