# lib/api-zod/src/generated/types/verifyCheckoutResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyCheckoutResponsePurchaseType } from "./verifyCheckoutResponsePurchaseType";

export interface VerifyCheckoutResponse {
  success: boolean;
  purchaseType: VerifyCheckoutResponsePurchaseType;
  artworkTitle: string;
  message: string;
}